backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
django-extensions==3.2.3
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
